perf: improve performance in create and explore sites by adding terser and splitting webpack config files the same way as in fast-website
