+++ sites/fast-component-explorer/package.json
@@ -5,12 +5,12 @@
   "sideEffects": false,
   "private": true,
   "scripts": {
-    "build:app": "webpack --progress --mode=production",
+    "build:app": "webpack --config=./webpack.prod.js",
     "convert:readme": "node ./build/convert-readme.js",
     "clean:tmp": "node ../../build/clean.js app/.tmp",
     "prettier": "prettier --config ../../.prettierrc --write \"**/*.{ts,tsx,html}\"",
     "prettier:diff": "prettier --config ../../.prettierrc \"**/*.{ts,tsx,html}\" --list-different",
-    "start": "webpack-dev-server",
+    "start": "webpack-dev-server --config=./webpack.dev.js --history-api-fallback --progress",
     "test": "yarn convert:readme && yarn eslint && yarn build:app",
     "eslint": "eslint . --ext .ts,.tsx",
     "eslint:fix": "eslint . --ext .ts,.tsx --fix"
@@ -45,7 +45,7 @@
     "babel-core": "^7.0.0-bridge.0",
     "babel-loader": "^8.0.6",
     "chroma-js": "^1.3.0",
-    "clean-webpack-plugin": "^1.0.0",
+    "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "^6.0.3",
     "css-loader": "^4.2.0",
     "csstype": "^2.5.8",
@@ -76,6 +76,7 @@
     "webpack-bundle-analyzer": "^3.0.3",
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.11.0",
+    "webpack-merge": "^4.2.2",
     "webpack-shell-plugin": "^0.5.0",
     "workbox-webpack-plugin": "^3.6.3",
     "worker-loader": "^2.0.0"
