+++ sites/fast-creator/package.json
@@ -5,10 +5,10 @@
   "sideEffects": false,
   "private": true,
   "scripts": {
-    "build": "webpack --progress --mode=production",
+    "build": "webpack --config=./webpack.prod.js",
     "prettier": "prettier --config ../../.prettierrc --write \"**/*.{ts,tsx,html}\"",
     "prettier:diff": "prettier --config ../../.prettierrc \"**/*.{ts,tsx,html}\" --list-different",
-    "start": "webpack-dev-server",
+    "start": "webpack-dev-server --config=./webpack.dev.js --history-api-fallback --progress",
     "test": "yarn eslint && yarn build",
     "eslint": "eslint . --ext .ts,.tsx",
     "eslint:fix": "eslint . --ext .ts,.tsx --fix"
@@ -36,7 +36,7 @@
     "@types/react-router": "^4.0.0",
     "babel-core": "^7.0.0-bridge.0",
     "babel-loader": "^8.0.6",
-    "clean-webpack-plugin": "^1.0.0",
+    "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "^6.0.3",
     "css-loader": "^4.2.0",
     "csstype": "^2.5.8",
@@ -57,6 +57,7 @@
     "react-dom": "^16.8.0",
     "react-router-dom": "^5.1.2",
     "style-loader": "^1.2.1",
+    "terser-webpack-plugin": "^4.1.0",
     "ts-jest": "^25.4.0",
     "ts-loader": "^7.0.2",
     "typescript": "^3.9.0",
@@ -64,6 +65,7 @@
     "webpack-bundle-analyzer": "^3.0.3",
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.11.0",
+    "webpack-merge": "^4.2.2",
     "workbox-webpack-plugin": "^3.6.3",
     "worker-loader": "^2.0.0"
   },
